Refactoring the entire SimpleCov parser and reports submition methodology (#3)
# SimpleCov parser refactor
The SimpleCov parser have been re-factored to the new API, renamed to a more fitting and suitable name, tests were added, support for multiple suit names have been implemented and coverage reports data models have been created to accommodate a better report structure.

In full detail:
* Fixed SimpleCovRcovJsonParser implementation
* Support for different types of Testing Frameworks (Rspec, MiniTest, Cucumber, etc.)
* Add data models for SimpleCov reports
* New JSON parsing functional
* Added Mark data model
* Added tests for data models
* Updated SimpleCov parser to new API
* Optimized SimpleCovRcovJson parser imports
* Added property for selecting coverage test suites
* Added unit tests for multiple suites usage cases
* Cleaned up CoverageReport data class
* Seperated and refactored `SimpleCovRcovJsonParserImpl` into it's different responsibilities(parsing and models)
* Renamed `SimpleCovRcovJsonParser` to `CoverageReportFileAnalyzer`

> Due to the extraction logic of parsing the JSON report file into separate classes,
the old class does not have any responsibility for parsing. So the old
name could be a little bit confusing. That's why a decision was made to rename it.

* Added documentation for `CoverageReportFileAnalyzer`

**Thank you @sergio1990 for this awesome work!!**